iroh_dns::attrs

Struct TxtAttrs

pub struct TxtAttrs<T> { /* private fields */ }

Attributes parsed from IROH_TXT_NAME TXT records.

This struct is generic over the key type. When using with String, this will parse all attributes. Can also be used with an enum, if it implements FromStr and Display.

Implementations

impl<T: FromStr + Display + Hash + Ord> TxtAttrs<T>

pub fn from_parts( endpoint_id: EndpointId, pairs: impl Iterator<Item = (T, String)>, ) -> Self

Creates TxtAttrs from an endpoint id and an iterator of key-value pairs.

pub fn from_strings( endpoint_id: EndpointId, strings: impl Iterator<Item = String>, ) -> Result<Self, ParseError>

Creates TxtAttrs from an endpoint id and an iterator of “{key}={value}” strings.

pub fn attrs(&self) -> &BTreeMap<T, Vec<String>>

Returns the parsed attributes.

pub fn endpoint_id(&self) -> EndpointId

Returns the endpoint id.

pub fn from_txt_lookup( name: String, lookup: impl Iterator<Item = impl Display>, ) -> Result<Self, ParseError>

Parses TXT record lookup results.

The name is the queried DNS name. The lookup iterator yields TXT record values that implement Display.

pub fn from_pkarr_signed_packet( packet: &SignedPacket, ) -> Result<Self, ParseError>

Parses a pkarr::SignedPacket.

pub fn to_txt_strings(&self) -> impl Iterator<Item = String> + '_

Converts to {key}={value} strings.

pub fn to_pkarr_signed_packet( &self, secret_key: &SecretKey, ttl: u32, ) -> Result<SignedPacket, EncodingError>

Creates a pkarr::SignedPacket

This constructs a DNS packet and signs it with a SecretKey.